Выделил логику показа карточки фильма в отдельный презентер, разбил компонент с попапом на несколько компонентов
