+++ src/view/popup/film-details-comments-wrap-view.js
@@ -1,0 +1,12 @@
+import AbstractView from '../abstract-view.js';
+
+const createFilmDetailsCommentsWrapTemplate = () => (
+  `<section class="film-details__comments-wrap">
+  </section>`
+);
+
+export default class FilmDetailsCommentsWrapView extends AbstractView {
+  get template() {
+    return createFilmDetailsCommentsWrapTemplate();
+  }
+}
