+++ src/view/popup/film-details-section-view.js
@@ -1,0 +1,12 @@
+import AbstractView from '../abstract-view.js';
+
+const createFilmDetailsSectionTemplate = () => (
+  `<section class="film-details">
+</section>`
+);
+
+export default class FilmDetailsSectionView extends AbstractView {
+  get template() {
+    return createFilmDetailsSectionTemplate();
+  }
+}
